# Rewrite ruby/setup-msys2-gcc