--- a/windows.js
+++ b/windows.js
@@ -1,42 +1,18 @@
 // 7z arguments
 //   -aoa overwrite existing, -bd disable progress indicator
 
-// OpenSSL version detection is needed in installGCCTools and installJRubyTools
-
 const fs = require('fs')
 const os = require('os')
 const path = require('path')
 const cp = require('child_process')
 const core = require('@actions/core')
 const exec = require('@actions/exec')
-const io = require('@actions/io')
 const tc = require('@actions/tool-cache')
 const common = require('./common')
 const rubyInstallerVersions = require('./windows-versions.json')
+const toolchainVersions = require('./windows-toolchain-versions.json')
 
-const drive = common.drive
-
-const msys2GCCReleaseURI  = 'https://github.com/ruby/setup-msys2-gcc/releases/download'
-
-const msys2BasePath = process.env['GHCUP_MSYS2'] || 'C:\\msys64'
-const vcPkgBasePath = process.env['VCPKG_INSTALLATION_ROOT']
-
-// needed for Ruby 2.0-2.3, cert file used by Git for Windows
-const certFile = [
-  'C:\\Program Files\\Git\\mingw64\\etc\\ssl\\cert.pem',
-  'C:\\Program Files\\Git\\mingw64\\ssl\\cert.pem',
-  'C:\\Program Files\\Git\\clangarm64\\etc\\ssl\\cert.pem',
-  'C:\\Program Files\\Git\\clangarm64\\ssl\\cert.pem'
-].find(file => fs.existsSync(file))
-if (certFile === undefined) {
-  throw new Error("Cannot find Git's cert file")
-}
-
-// location & path for old RubyInstaller DevKit (MSYS1), used with Ruby 2.0-2.3
-const msys1 = `${drive}:\\DevKit64`
-const msysPathEntries = [`${msys1}\\mingw\\x86_64-w64-mingw32\\bin`, `${msys1}\\mingw\\bin`, `${msys1}\\bin`]
-
-export function getAvailableVersions(platform, engine) {
+export function getAvailableVersions(_platform, engine) {
   if (engine === 'ruby') {
     return Object.keys(rubyInstallerVersions).filter(version => os.arch() in rubyInstallerVersions[version])
   } else {
@@ -47,14 +23,10 @@ export function getAvailableVersions(platform, engine) {
 export async function install(platform, engine, version) {
   const url = rubyInstallerVersions[version][os.arch()]
 
-  // The windows-2016 and windows-2019 images have MSYS2 build tools (C:/msys64/usr)
-  // and MinGW build tools installed.  The windows-2022 image has neither.
-  const hasMSYS2PreInstalled = ['windows-2019', 'windows-2016'].includes(platform)
-
   if (!url.endsWith('.7z')) {
     throw new Error(`URL should end in .7z: ${url}`)
   }
-  const base = url.slice(url.lastIndexOf('/') + 1, url.length - '.7z'.length)
+  const base = path.posix.basename(url, '.7z')
 
   let rubyPrefix, inToolCache
   if (common.shouldUseToolCache(engine, version)) {
@@ -65,121 +37,68 @@ export async function install(platform, engine, version) {
       rubyPrefix = common.getToolCacheRubyPrefix(platform, engine, version)
     }
   } else {
-    rubyPrefix = `${drive}:\\${base}`
+    rubyPrefix = `${common.drive}:\\${base}`
   }
 
   if (!inToolCache) {
     await downloadAndExtract(engine, version, url, base, rubyPrefix);
   }
 
+  const paths = [`${rubyPrefix}\\bin`]
   const windowsToolchain = core.getInput('windows-toolchain')
   if (windowsToolchain === 'none') {
-    common.setupPath([`${rubyPrefix}\\bin`])
+    common.setupPath(paths)
     return rubyPrefix
   }
 
-  let toolchainPaths = (version === 'mswin') ? await setupMSWin() : await setupMingw(version)
-  const msys2Type = common.setupPath([`${rubyPrefix}\\bin`, ...toolchainPaths])
+  const toolchainUrl = toolchainVersions[version][os.arch()]
 
-  // install msys2 tools for all Ruby versions, only install mingw or ucrt for Rubies >= 2.4
+  // Examples:
+  // - DevKit-mingw64-64-4.7.2-20130224-1432-sfx.exe
+  // - msys2-clangarm64.7z
+  // - vcpkg-x64-windows.7z
+  const toolchainName = path.posix.basename(toolchainUrl).split('-')[0].toLowerCase()
 
-  if (!hasMSYS2PreInstalled) {
-    await installMSYS2Tools()
+  if (toolchainName === 'msys2') {
+    paths.push(...await installMSYS2(toolchainUrl, rubyPrefix))
+  } else {
+    if (toolchainName === 'vcpkg') {
+      paths.push(...await installVCPKG(toolchainUrl))
+    } else if (toolchainName === 'devkit') {
+      paths.push(...await installMSYS1(toolchainUrl))
+    } else {
+      throw new Error(`Unknown toolchain type: ${toolchainUrl}`)
+    }
+
+    // Install msys2 for other rubies to provide better command line tools
+    paths.push(...await installMSYS2(toolchainVersions['head'][os.arch()]))
   }
 
-  // windows 2016 and 2019 need ucrt64 installed, 2022 and future images need
-  // ucrt64 or mingw64 installed, depending on Ruby version
-  if ((msys2Type === 'ucrt64') || (common.floatVersion(version) >= 2.4)) {
-    await installGCCTools(msys2Type, version)
-  }
-
-  if (version === 'mswin') {
-    await installVCPkg()
-  }
-
-  const ridk = `${rubyPrefix}\\bin\\ridk.cmd`
-  if (fs.existsSync(ridk)) {
-    await common.measure('Adding ridk env variables', async () => addRidkEnv(ridk))
-  }
-
+  common.setupPath(paths)
   return rubyPrefix
 }
 
-// Actions windows-2022 image does not contain any mingw or ucrt build tools.  Install tools for it,
-// and also install ucrt tools on earlier versions, which have msys2 and mingw tools preinstalled.
-async function installGCCTools(type, version) {
-  // 2022-Dec ruby/msys2-gcc-pkgs now uses a suffix to delineate different archive versions.
-  // At present, the only use is to indicate the included OpenSSL version.
-  // With no suffix, archives include OpenSSL 1.1.1, with a '-3.0' suffix, they include
-  // OpenSSL 3.0.x.  Note that the mswin archive uses OpenSSL 3.
-  // As of Jan-2023, OpenSSL 3 is used in Ruby 3.2 & all head versions.  MSYS2 updated
-  // to OpenSSL 3 on 14-Jan-2023.
-  let suffix = ''
-  if (common.isHeadVersion(version) || common.floatVersion(version) >= 3.2) {
-    suffix = '-3.0'
-  }
-  const downloadPath = await common.measure(`Downloading ${type}${suffix} build tools`, async () => {
-    let url = `${msys2GCCReleaseURI}/msys2-gcc-pkgs/${type}${suffix}.7z`
-    console.log(url)
-    return await tc.downloadTool(url)
-  })
-
-  await common.measure(`Extracting  ${type}${suffix} build tools`, async () =>
-    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${msys2BasePath}`], { silent: true }))
-}
-
-// Actions windows-2022 image does not contain any MSYS2 build tools.  Install tools for it.
-// A subset of the MSYS2 base-devel group
-async function installMSYS2Tools() {
-  const downloadPath = await common.measure(`Downloading msys2 build tools`, async () => {
-    const suffix = os.arch() === 'arm64' ? '-arm64' : ''
-
-    let url = `${msys2GCCReleaseURI}/msys2-gcc-pkgs/msys2${suffix}.7z`
-    console.log(url)
-    return await tc.downloadTool(url)
-  })
-
-  // need to remove all directories, since they may indicate old packages are installed,
-  // otherwise, error of "error: duplicated database entry"
-  fs.rmSync(`${msys2BasePath}\\var\\lib\\pacman\\local`, { recursive: true, force: true })
-
-  await common.measure(`Extracting  msys2 build tools`, async () =>
-    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${msys2BasePath}`], { silent: true }))
-}
-
-// Windows JRuby can install gems that require compile tools, only needed for
-// windows-2022 and later images
-export async function installJRubyTools() {
-  await installMSYS2Tools()
-  await installGCCTools(os.arch() === 'arm64' ? 'clangarm64' : 'mingw64', '3.4')
-}
-
-// Install vcpkg files needed to build mswin Ruby
-async function installVCPkg() {
-  const downloadPath = await common.measure(`Downloading mswin vcpkg packages`, async () => {
-    let url = `${msys2GCCReleaseURI}/msys2-gcc-pkgs/mswin.7z`
-    console.log(url)
-    return await tc.downloadTool(url)
-  })
-
-  await common.measure(`Extracting  mswin vcpkg packages`, async () =>
-    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${vcPkgBasePath}`], { silent: true }))
-}
-
 async function downloadAndExtract(engine, version, url, base, rubyPrefix) {
-  const parentDir = path.dirname(rubyPrefix)
-
   const downloadPath = await common.measure('Downloading Ruby', async () => {
     console.log(url)
     return await tc.downloadTool(url)
   })
 
+  const extractPath = process.env.RUNNER_TEMP
   await common.measure('Extracting  Ruby', async () =>
     // -xr extract but exclude share\doc files
-    exec.exec('7z', ['x', downloadPath, '-bd', `-xr!${base}\\share\\doc`, `-o${parentDir}`], { silent: true }))
+    exec.exec('7z', ['x', downloadPath, '-bd', `-xr!${base}\\share\\doc`, `-o${extractPath}`], { silent: true }))
 
-  if (base !== path.basename(rubyPrefix)) {
-    await io.mv(path.join(parentDir, base), rubyPrefix)
+  const parentDir = path.dirname(rubyPrefix)
+  if (!fs.existsSync(parentDir)) {
+    fs.mkdirSync(parentDir, { recursive: true })
+  }
+
+  const extractedPath = path.join(extractPath, base)
+  if (extractedPath[0] === rubyPrefix[0]) {
+    fs.renameSync(extractedPath, rubyPrefix)
+  } else {
+    fs.symlinkSync(extractedPath, rubyPrefix, 'junction')
   }
 
   if (common.shouldUseToolCache(engine, version)) {
@@ -187,108 +106,122 @@ async function downloadAndExtract(engine, version, url, base, rubyPrefix) {
   }
 }
 
-async function setupMingw(version) {
-  core.exportVariable('MAKE', 'make.exe')
-
-  // rename these to avoid confusion when Ruby is using OpenSSL 1.0.2.
-  // most current extconf files look for 1.1.x dll files first, which is the version of the renamed files
-  if (common.floatVersion(version) <= 2.4) {
-    renameSystem32Dlls()
-  }
-
-  if (common.floatVersion(version) <= 2.3) {
-    core.exportVariable('SSL_CERT_FILE', certFile)
-    await common.measure('Installing MSYS1', async () => installMSYS1(version))
-    return msysPathEntries
-  } else {
-    return []
-  }
+export async function installJRubyTools() {
+  return await installMSYS2(toolchainVersions['head'][os.arch()])
 }
 
-// Ruby 2.0-2.3
-async function installMSYS1(version) {
-  const url = 'https://github.com/oneclick/rubyinstaller/releases/download/devkit-4.7.2/DevKit-mingw64-64-4.7.2-20130224-1432-sfx.exe'
-  const downloadPath = await tc.downloadTool(url)
-  await exec.exec('7z', ['x', downloadPath, `-o${msys1}`], { silent: true })
+async function installMSYS2(url, rubyPrefix = process.env.RUNNER_TEMP) {
+  const downloadPath = await common.measure('Downloading msys2 build tools', async () => {
+    console.log(url)
+    return await tc.downloadTool(url)
+  })
 
-  // below are set in the old devkit.rb file ?
-  core.exportVariable('RI_DEVKIT', msys1)
-  core.exportVariable('CC' , 'gcc')
-  core.exportVariable('CXX', 'g++')
-  core.exportVariable('CPP', 'cpp')
-  core.info(`Installed RubyInstaller DevKit for Ruby ${version}`)
+  const extractPath = path.join(process.env.RUNNER_TEMP, 'msys64')
+  await common.measure('Extracting  msys2 build tools', async () =>
+    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${extractPath}`], { silent: true }))
+
+  const msys2Path = path.join(rubyPrefix, 'msys64')
+  if (extractPath !== msys2Path) {
+    fs.symlinkSync(extractPath, msys2Path, 'junction')
+  }
+
+  const ridk = `${rubyPrefix}\\bin\\ridk.cmd`
+  if (fs.existsSync(ridk)) {
+    await common.measure('Setting up ridk environment', async () => exportCommandEnv(`set MAKE=make && "${ridk}" enable`))
+  }
+
+  // Examples:
+  // - msys2-clangarm64.7z
+  // - msys2-ucrt64-gcc@14.7z
+  // - msys2-mingw64-gcc@14-openssl@1.1.7z
+  const msys2Type = path.posix.basename(url).split(/[-.]/)[1]
+  return [`${msys2Path}\\${msys2Type}\\bin`, `${msys2Path}\\usr\\bin`]
 }
 
-async function setupMSWin() {
-  core.exportVariable('MAKE', 'nmake.exe')
-  return await common.measure('Setting up MSVC environment', async () => addVCVARSEnv())
+async function installMSYS1(url) {
+  const certFile = [
+    'C:\\Program Files\\Git\\mingw64\\etc\\ssl\\cert.pem',
+    'C:\\Program Files\\Git\\mingw64\\ssl\\cert.pem',
+  ].find(file => fs.existsSync(file))
+  if (certFile === undefined) {
+    throw new Error("Cannot find Git's cert file")
+  }
+
+  const downloadPath = await common.measure('Downloading msys1 build tools', async () => {
+    console.log(url)
+    return await tc.downloadTool(url)
+  })
+
+  const msys1Path = `${common.drive}:\\DevKit64`
+  await common.measure('Extracting  msys1 build tools', async () =>
+    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${msys1Path}`], { silent: true }))
+
+  await common.measure('Setting up DevKit environment', async () => {
+    // ssl dlls cause issues with msys1 Rubies
+    const sys32 = 'C:\\Windows\\System32\\'
+    const badFiles = [`${sys32}libcrypto-1_1-x64.dll`, `${sys32}libssl-1_1-x64.dll`]
+    const existing = badFiles.filter((dll) => fs.existsSync(dll))
+    if (existing.length > 0) {
+      core.warning(`Renaming ${existing.join(' and ')} to avoid dll resolution conflicts on Ruby < 2.4`)
+      existing.forEach(dll => fs.renameSync(dll, `${dll}_`))
+    }
+
+    exportEnv({
+      CC: 'gcc',
+      CPP: 'cpp',
+      CXX: 'g++',
+      MAKE: 'make',
+      RI_DEVKIT: msys1Path,
+      SSL_CERT_FILE: certFile
+    })
+  })
+
+  return [`${msys1Path}\\mingw\\x86_64-w64-mingw32\\bin`, `${msys1Path}\\mingw\\bin`, `${msys1Path}\\bin`]
 }
 
 /* Sets MSVC environment for use in Actions
  *   allows steps to run without running vcvars*.bat, also for PowerShell
- *   adds a convenience VCVARS environment variable
  *   this assumes a single Visual Studio version being available in the Windows images */
-export function addVCVARSEnv() {
-  let cmd = 'vswhere -latest -property installationPath'
-  let vcVars = `${cp.execSync(cmd).toString().trim()}\\VC\\Auxiliary\\Build\\vcvars64.bat`
-
-  if (!fs.existsSync(vcVars)) {
-    throw new Error(`Missing vcVars file: ${vcVars}`)
-  }
-  core.exportVariable('VCVARS', vcVars)
-
-  cmd = `cmd.exe /c ""${vcVars}" && set"`
-
-  let newEnv = new Map()
-  let newSet = cp.execSync(cmd).toString().trim().split(/\r?\n/)
-  newSet = newSet.filter(line => /\S=\S/.test(line))
-  newSet.forEach(s => {
-    let [k,v] = common.partition(s, '=')
-    newEnv.set(k,v)
+async function installVCPKG(url) {
+  const downloadPath = await common.measure('Downloading mswin vcpkg packages', async () => {
+    console.log(url)
+    return await tc.downloadTool(url)
   })
 
-  let newPathEntries = undefined
-  for (let [k, v] of newEnv) {
-    if (process.env[k] !== v) {
-      if (/^Path$/i.test(k)) {
-        const newPathStr = v.replace(`${path.delimiter}${process.env['Path']}`, '')
-        newPathEntries = newPathStr.split(path.delimiter)
-      } else {
-        core.exportVariable(k, v)
-      }
+  const extractPath = process.env.VCPKG_INSTALLATION_ROOT
+  await common.measure('Extracting  mswin vcpkg packages', async () =>
+    exec.exec('7z', ['x', downloadPath, '-aoa', '-bd', `-o${extractPath}`], { silent: true }))
+
+  return await common.measure('Setting up MSVC environment', async () => {
+    const cmd = 'vswhere -latest -property installationPath'
+    const vcVarsBat = os.arch() === 'arm64' ? 'vcvarsarm64.bat' : 'vcvars64.bat'
+    const vcVars = `${cp.execSync(cmd).toString().trim()}\\VC\\Auxiliary\\Build\\${vcVarsBat}`
+
+    if (!fs.existsSync(vcVars)) {
+      throw new Error(`Missing vcVars file: ${vcVars}`)
     }
-  }
-  return newPathEntries
-}
 
-// ssl files cause issues with non RI2 Rubies (<2.4) and ruby/ruby's CI from build folder due to dll resolution
-function renameSystem32Dlls() {
-  const sys32 = 'C:\\Windows\\System32\\'
-  const badFiles = [`${sys32}libcrypto-1_1-x64.dll`, `${sys32}libssl-1_1-x64.dll`]
-  const existing = badFiles.filter((dll) => fs.existsSync(dll))
-  if (existing.length > 0) {
-    console.log(`Renaming ${existing.join(' and ')} to avoid dll resolution conflicts on Ruby <= 2.4`)
-    existing.forEach(dll => fs.renameSync(dll, `${dll}_`))
-  }
-}
-
-// Sets MSYS2 ENV variables set from running `ridk enable`
-function addRidkEnv(ridk) {
-  let newEnv = new Map()
-  let cmd = `cmd.exe /c "${ridk} enable && set"`
-  let newSet = cp.execSync(cmd).toString().trim().split(/\r?\n/)
-  newSet = newSet.filter(line => /^\S+=\S+/.test(line))
-  newSet.forEach(s => {
-    let [k, v] = common.partition(s, '=')
-    newEnv.set(k, v)
+    return exportCommandEnv(`set MAKE=nmake && "${vcVars}"`)
   })
+}
 
-  for (let [k, v] of newEnv) {
+// Run a cmd command, export its environment, and return new paths
+function exportCommandEnv(command) {
+  const cmd = `cmd.exe /s /c "${command} >NUL && "${process.execPath}" -e console.log(JSON.stringify(process.env))"`
+  const env = JSON.parse(cp.execSync(cmd).toString())
+
+  const paths = env.Path.replace(`${path.delimiter}${process.env.Path}`, '').split(path.delimiter)
+  delete env.Path
+
+  exportEnv(env)
+  return paths
+}
+
+function exportEnv(env) {
+  for (const [k, v] of Object.entries(env)) {
     if (process.env[k] !== v) {
-      if (!/^Path$/i.test(k)) {
-        console.log(`${k}=${v}`)
-        core.exportVariable(k, v)
-      }
+      console.log(`${k}=${v}`)
+      core.exportVariable(k, v)
     }
   }
 }