… .github/workflows/test.yml
… dist/index.js
… windows-toolchain-versions.json