This PR include a few changes for windows:

- Fix the issue with gcc 15/14 and mingw-w64 r688/r720 by introducing flexible variants of toolchains.
- Reduce friction with https://github.com/ruby/setup-msys2-gcc/pull/26.
- Improves performance by install to fast disk and create junction link on slow disk.
    - Ruby is installed to `$RUNNER_TEMP\$base` and a junction link is created at original `$rubyPrefix` if `$rubyPrefix` is on the slow disk.
    - MSYS2 is installed to `$RUNNER_TEMP\msys64` and a junction link is created at `$rubyPrefix\msys64`.
- Improves the reliability of `ridk enable` step by using proper serialization.
- Refactor and clean up tech debt.

### Performance

Run to run improvement varies a lot due to different sizes of packages and the unstable performance of the slow disk.
However, in general this saves anywhere from 20s to 60s per run. Example of a performance improvement from 63.43s to 27.15s - or 36.38s in saving:

https://github.com/ruby/setup-ruby/actions/runs/14772696897/job/41475457782

```
Downloading Ruby
  https://github.com/oneclick/rubyinstaller2/releases/download/RubyInstaller-3.4.3-1/rubyinstaller-3.4.3-1-x64.7z
  Took   0.97 seconds
Extracting  Ruby
  Took  25.38 seconds
Modifying Path
Downloading msys2 build tools
  https://github.com/ruby/setup-msys2-gcc/releases/download/msys2-gcc-pkgs/msys2.7z
  Took   0.73 seconds
Extracting  msys2 build tools
  Took  12.52 seconds
Downloading ucrt64-3.0 build tools
  https://github.com/ruby/setup-msys2-gcc/releases/download/msys2-gcc-pkgs/ucrt64-3.0.7z
  Took   1.00 seconds
Extracting  ucrt64-3.0 build tools
  Took  22.83 seconds
```

https://github.com/ntkme/setup-ruby/actions/runs/14823150160/job/41612838924

```
Downloading Ruby
  https://github.com/oneclick/rubyinstaller2/releases/download/RubyInstaller-3.4.3-1/rubyinstaller-3.4.3-1-x64.7z
  Took   0.46 seconds
Extracting  Ruby
  Took   9.87 seconds
Modifying Path
Downloading msys2 build tools
  https://github.com/ntkme/setup-msys2-gcc/releases/latest/download/msys2-ucrt64-gcc@14.7z
  Took   1.18 seconds
Extracting  msys2 build tools
  Took  15.64 seconds
```

---

Note regarding why use `$rubyPrefix\msys64`:
- `ENV["MSYS2_PATH"]` have highest priority, but it's only supported on ruby >=3.0.7: https://github.com/oneclick/rubyinstaller2/commit/a4590dd112705a9b1042d5d03fb253c8b6a5b669 
- `$rubyPrefix\msys64` has the second highest priority, and it's supported by ruby >=2.4.1: https://github.com/oneclick/rubyinstaller2/commit/21a416ffb73c876a344f08f2d362b5a381984b12

`>=2.4.1` means `$rubyPrefix\msys64` is supported by all currently available ruby installer versions that's linked with msys2, and it has higher priority than `C:\msys64` so that we don’t need to touch the preinstalled `C:\msys64` at all.