@MSP-Greg 2025-05-12
Curious, why the need for 'Revert back to use `GITHUB_WORKSPACE` to detect fast drive'?