@MSP-Greg 2025-05-13
@eregon

I've quickly looked as this, and everything seems ok.  We're getting close to finalizing the changes to `setup-msys2-gcc`, so checking is this is okay.  Once the new packages are created, this can be merged.