@eregon 2025-05-14
From the PR title I thought this was a ruby/setup-msys2-gcc PR :sweat_smile: 
I'll try to take a look soon.