@ntkme 2025-05-14
clangarm64 ruby head test is now passing after I disabled the "update bundler" step in setup-ruby. The problem is that in this repo we're intentionally setting up msys2 after installing ruby during our test. Not having msys2 installed currently causes the built-in `date` gem to crash, and the "update bundler" step loads the `date` gem.

See https://github.com/oneclick/rubyinstaller2-packages/issues/25#issuecomment-2881729508 and https://github.com/ruby/date/issues/126#issuecomment-2881751295