@MSP-Greg 2025-05-15
@eregon

> I guess the order is merge https://github.com/ruby/setup-msys2-gcc/pull/26, create the release & assets, then update this PR to use ruby/setup-msys2-gcc links, and merge this PR?

The PR has been merged in https://github.com/ruby/setup-msys2-gcc