@ntkme 2025-05-15
@eregon Release artifacts has been uploaded and I've updated this PR to use ruby/setup-msys2-gcc. There is a truffleruby test that's failing but it's unrelated to my change.