@ronaldtse 2025-05-16
Hoping this gets merged soon as we have a number of release pipelines blocked due to the recent Ruby 3.1 on Windows issue. Thank you for the excellent work and maintenance! 🙇‍♂️