@eregon 2025-05-28
A bit more debugging in https://github.com/eregon/actions-shell/actions/runs/15309439156/job/43070305903#step:4:550 shows `D:/a/_temp/` paths in $" but `C:/hostedtoolcache/windows/Ruby/3.4.4/x64/` in $:, which seems not unlikely to cause some incompatibility in gem test suites.
And indeed `File.realpath("C:/hostedtoolcache/windows/Ruby/3.4.4/x64/lib/ruby/site_ruby/3.4.0") = "D:/a/_temp/rubyinstaller-3.4.4-2-x64/lib/ruby/site_ruby/3.4.0"`