@ntkme 2025-05-28
> which seems not unlikely to cause some incompatibility in gem test suites.

Yes, I understand the junction can cause issues when some code use a mix of `realpath` vs `absolute_path`. The main reason to use junction is to let us extract files on the faster disk.

I wish the code run into the issue can be more consistent in that either to use `realpath` everywhere or use `absolute_path` everywhere but never a mix of the two, but not sure if that's a fair ask. If we want to pursue better compatibility than performance, we can easily revert the usage of junction and just extract to slow disk as before.