@MSP-Greg 2025-05-28
@eregon

IIRC, Ruby head builds have always used D:\ for their `PATH` addition, so it should work fine.  Rubies that are part of the runner image will need to stay in `C:/hostedtoolcache/windows/Ruby`.

I probably can't get to it tonite, tomorrow should be better...

Lastly, I also recall that there are two kinds of Windows links, may look at that also.