@ntkme 2025-05-28
> Lastly, I also recall that there are two kinds of Windows links, may look at that also.

Correct, there are two kinds of soft links: junction and symlink. There is also hard link but hard link can not have directory as target.

There is a good explanation about the difference here: https://superuser.com/questions/343074/directory-junction-vs-directory-symbolic-link.

The biggest concern I had at the time was that creating symlink require privileged access, while junction only need regular access.