@ntkme 2025-05-28
Another idea here:

Detect if we're in self-hosted runner - If so extract into hosttoolcache (no junction) directly, otherwise extract into `$RUNNER_TEMP` (also no junction) directly.