@eregon 2025-05-28
> Detect if we're in self-hosted runner - If so extract into hosttoolcache (no junction) directly, otherwise extract into `$RUNNER_TEMP` (also no junction) directly.

It should work, but since no one complained and probably no one uses self-hosted runners on Windows with persistent toolcache let's keep it simple and do only that latter part always.
FYI detecting self-hosted runner is unfortunately not so reliable unfortunately.