@ntkme 2025-05-28
> FYI detecting self-hosted runner is unfortunately not so reliable unfortunately.

https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/store-information-in-variables#default-environment-variables

> RUNNER_ENVIRONMENT

> The environment of the runner executing the job. Possible values are: `github-hosted` for GitHub-hosted runners provided by GitHub, and self-hosted for `self-hosted` runners configured by the repository owner.