@eregon 2025-05-28
Interesting, but in reality it's more complicated than that as there are some self-hosted runners for which we can reuse prebuilt Rubies and some self-hosted runners where we can't. Anyway, simple is almost always better unless there is a need for something more complicated (I don't see one).